refactor ChildProcessHelper into a separate file
