+++ src/electron/sslibev_badvpn_tunnel.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ChildProcess, execSync, spawn} from 'child_process';
+import {execSync} from 'child_process';
 import {powerMonitor} from 'electron';
 import {platform} from 'os';
 import * as path from 'path';
@@ -23,6 +23,7 @@ import * as errors from '../www/model/errors';
 import {ShadowsocksConfig} from '../www/model/shadowsocks';
 
 import {checkUdpForwardingEnabled, isServerReachable, validateServerCredentials} from './connectivity';
+import {ChildProcessHelper} from './process';
 import {RoutingDaemon} from './routing_service';
 import {pathToEmbeddedBinary} from './util';
 import {VpnTunnel} from './vpn_tunnel';
@@ -285,76 +286,6 @@ export class ShadowsocksLibevBadvpnTunnel implements VpnTunnel {
   }
 }
 
-// Simple "one shot" child process launcher.
-//
-// NOTE: Because there is no way in Node.js to tell whether a process launched successfully,
-//       #startInternal always succeeds; use #onExit to be notified when the process has exited
-//       (which may be immediately after calling #startInternal if, e.g. the binary cannot be
-//       found).
-class ChildProcessHelper {
-  private process?: ChildProcess;
-  protected isInDebugMode = false;
-
-  private exitListener?: () => void;
-
-  protected constructor(private path: string) {}
-
-  /**
-   * Starts the process with the given args. If enableDebug() has been called, then the process is started in verbose mode if supported.
-   * @param args The args for the process
-   */
-  protected launch(args: string[]) {
-    this.process = spawn(this.path, args);
-    const processName = path.basename(this.path);
-
-    const onExit = (code: number, signal: string) => {
-      if (this.process) {
-        this.process.removeAllListeners();
-      }
-      if (this.exitListener) {
-        this.exitListener();
-      }
-
-      logExit(processName, code, signal);
-    };
-
-    if (this.isInDebugMode) {
-      // Expose logs to the node output.  This also makes the logs available in Sentry.
-      this.process.stdout.on('data', (data) => console.log(`[STDOUT - ${processName}]: ${data}`));
-      this.process.stderr.on('data', (data) => console.error(`[STDERR - ${processName}]: ${data}`));
-    }
-
-    // We have to listen for both events: error means the process could not be launched and in that
-    // case exit will not be invoked.
-    this.process.on('error', onExit.bind((this)));
-    this.process.on('exit', onExit.bind((this)));
-  }
-
-  // Use #onExit to be notified when the process exits.
-  stop() {
-    if (!this.process) {
-      // Never started.
-      if (this.exitListener) {
-        this.exitListener();
-      }
-      return;
-    }
-
-    this.process.kill();
-  }
-
-  set onExit(newListener: (() => void)|undefined) {
-    this.exitListener = newListener;
-  }
-
-  /**
-   * Enables verbose logging for the process.  Must be called before launch().
-   */
-  public enableDebugMode() {
-    this.isInDebugMode = true;
-  }
-}
-
 class SsLocal extends ChildProcessHelper {
   constructor(private readonly proxyPort: number) {
     super(pathToEmbeddedBinary('shadowsocks-libev', 'ss-local'));
@@ -405,20 +336,5 @@ class Tun2socks extends ChildProcessHelper {
     args.push('--loglevel', this.isInDebugMode ? 'info' : 'error');
 
     this.launch(args);
-  }
-}
-
-function logExit(processName: string, exitCode?: number, signal?: string) {
-  const prefix = `[EXIT - ${processName}]: `;
-  if (exitCode !== null) {
-    const log = exitCode === 0 ? console.log : console.error;
-    log(`${prefix}Exited with code ${exitCode}`);
-  } else if (signal !== null) {
-    const log = signal === 'SIGTERM' ? console.log : console.error;
-    log(`${prefix}Killed by signal ${signal}`);
-  } else {
-    // This should never happen.  It likely signals an internal error in Node, as it is supposed to
-    // always pass either an exit code or an exit signal to the exit handler.
-    console.error(`${prefix}Process exited for an unknown reason.`);
   }
 }
